Upgrades Alert to Stimulus V3
Adds new functionality that's backward compatible with code prior to this PR.

Now, you can pass a 0 (zero) value to dismissAfterValue, which will behave the same as
not supplying the attribute.  The rational for this change is it's easier to write
cleaner code in the html.erb to handle the "make the user close this flash" use case
when you can supply a value.

So, this:
<div data-controller='alert'
     <% if alert_dismiss_value.present? %>
        data-alert-dismiss-after-value=<%= alert_dismiss_value %>
     <% end %>

becomes:
<div data-controller='alert'
     data-alert-dismiss-after-value=<%= alert_dismiss_value %>

which feels cleaner

Also, adds some defensive code around adding and removing the ShowClass
and HideClass, just in case someone wants to deliver that functionality
externally to this controller.