Fixes open issue #154 -- enter animation isn't working
Acording to the open issue, the root cause is trying animate after removing 'hidden'
from the menu target.  I wish I knew CSS well enough to know if this root cause analysis
is correct.

Regardless, we throw in a strategic 'invisible' and remove 'hidden' early in the show().
Later, we remove the 'invisible' as if it was the original 'hidden', which does seem to
fix the show animation problem.