Adds the ability to dynamically add / remove click@window->dropdown#hide
This one is debatable:  I hate, hate, hate the clutter of the click@window->... events firing on different controllers when debugging.
So, this adds two ways to turn trigger the new functionality: remove-window-action-value(boolean) and windowActionTargets.

Either way, this code will look for a buttonTarget (where most of the time, all the action happens) and
windowActionTargets (just in case someone has click handlers on more than the buttonTarget) and dynamically
removes the click@window->dropdown#hide handler when the dropdown is closed and adds it back when the dropdown is open.

I also added some comments to the code for future code spelunkers.